smap: actc_t 收窄为 u8 并以 sqrt 压缩高频频次
删除 addr 字段后普通虚机场景走 SeparateStrategy 桶排序(桶规格 256),
内核态扫描 freq 可超 256, 超 255 的页全压进 255 桶致频繁次失区分。
actc_t 收窄为 u8(仅导出/传输类型)与 FREQ_BUCKETS_SIZE 对齐。

- access_bit_actc_data 累加数组跟随 actc_t 保留 u8: 软件扫描频次由
  扫描/迁移周期决定、一般不高, u8 不回绕; 软件路径(accessed_bit.c ++,
  access_pid.c fill_actc_data_by_bitmap 读出)不压缩, u8 直接用, 保
  低频区分度。
- 仅硬件扫描路径做 u16->u8 压缩: hist_ops.c do_actc_update 把 BA 扫描
  u16 频次经 compress_freq 开根号压成 u8 后累加进 access_bit_actc_data,
  钳 U8_MAX; compress_freq 现唯一源码使用者即 do_actc_update。
- compress_freq(raw)=floor(sqrt(raw)): 逐位试凑自包含, 不依赖内核
  int_sqrt(UT 桩环境无该符号); u16 输入结果恒 <=255, 单调无断崖。
- HAM 频次保留 u16 不压缩: accessed_bit.h struct freq_info.freq 改回
  u16, 与 ham_migration.h/ham_tasks_mgr.h(ham_page_map.freq=unsigned
  short)对齐避 cross-module 布局错位; get_freq_info 直接赋原始 u16。
- GET_TRACKING 为 DFX 统计扫描频次, 保留原始 u16 不压缩:
  tracking_info_payload.data -> u16*, u16 累加后直接 copy_to_user;
  用户态 TrakingInfoPayload.data -> uint16_t* 同步。
- hist 临时 hot-window buffer(tmpbuffer/copy_actc_to_buf dst_buf)保留
  u16 不压缩: 仅供 calc_4k_scan_hot_wins max 选举, 不进
  access_bit_actc_data, 不导出; addr_seg.max 本就 u16 无损承接。
- 用户态联动: ActCount.freqMin/freqMax 改 u8; separate_strategy
  l1FreqMax/l2FreqMax 改 actc_t; 公开 API ubturbo_smap_freq_query 的
  uint16_t* 输出缓冲更宽, 0..255 无损, 不破外部 ABI。
- actc_data(packed) 3B->2B 内核/用户同步; freq_info 因 u64 对齐仍 16B。
- UT: access_bit_actc_data alloc 改 actc_t; do_actc_update 断言按
  compress 后值、overflow 钳 U8_MAX; 新增 CompressFreqSqrt 单测。

+++ plugins/smap/src/user/manage/manage.c
@@ -1790,7 +1790,7 @@ static void CalcActcStats(ProcessAttr *attr)
         }
 
         count->freqMax = 0;
-        count->freqMin = UINT16_MAX;
+        count->freqMin = UINT8_MAX;
         count->freqNum = 0;
         count->freqSum = 0;
         count->remoteHotNum = 0;
